+++ src/controllers/AdminController.php
@@ -33,6 +33,7 @@ class AdminController extends Controller
 		if ( ! is_null($this->layout))
 		{
 			$this->layout = View::make($this->layout);
+			$this->layout->page = false;
 		}
 	}
 
@@ -41,12 +42,11 @@ class AdminController extends Controller
 	 *
 	 * @return Response
 	 */
-	public function index()
+	public function index($modelName)
 	{
 		//set the layout content and title
 		$this->layout->content = View::make("administrator::index");
 	}
-
 
 	/**
 	 * Gets the item edit page / information
@@ -446,9 +446,23 @@ class AdminController extends Controller
 
 		//get the inputted rows and the model rows
 		$rows = (int) Input::get('rows', 20);
-		$dataTable->setRowsPerPage(App::make('session'), 0, $rows);
+		$dataTable->setRowsPerPage(App::make('session.store'), 0, $rows);
 
 		return Response::JSON(array('success' => true));
+	}
+
+	/**
+	 * The pages view
+	 *
+	 * @return Response
+	 */
+	public function page($page)
+	{
+		//set the page
+		$this->layout->page = $page;
+
+		//set the layout content and title
+		$this->layout->content = View::make($page);
 	}
 
 	/**