+++ src/lang/ca/knockout.php
@@ -1,0 +1,24 @@
+<?php 
+
+return array(
+
+	/*
+	|--------------------------------------------------------------------------
+	| Administrator Language Lines for knockout
+	|--------------------------------------------------------------------------
+	|
+	| ca - Catalan
+	|
+	*/
+
+	'delete_active_item' => 'Està segur que desitja eliminar aquest element? Aquesta operació és irreversible.',
+	'saving' => 'Guardant...',
+	'saved' => 'Element guardat.',
+	'deleting' => 'Eliminant...',
+	'deleted' => 'Element eliminat.',
+	'character_left' => ' caràcter restant',
+	'characters_left' => ' caràcters restants',
+	'no_results' => 'No s\'han trobat resultats',
+	'select_options' => 'Seleccioni Algunes Opcions',
+
+);